Created a separate file just for checking data
I focused on cleaning up my code, making it easier to read and add to. The code was becoming a little messy with so many comments and commented methods I was not using.

My solution was to create the separate file just for checking the data. This way the main file can be used for cleaning.

Ran into a small issue with the 'revenue' column. It is not reading and converting the value as I intended.

I will do further reading into .replace(), I think I am missing something small.