Updated README and more data cleaning
Continued with data cleaning, renaming columns and removing irrelevant columns that are not relevant for data analysis.

Made comprehensive updates to the README file. It now provides detailed information about the goal, process and results of the project. I will continue to update it as I move forward with project.